Closer to multichunk result combining
Fix subsample idx
Move specialist matrix functions out of matrix_ops and into gpu_api
Correctly map query vs ref chunk results
