+++ src/gpu_api.cpp
@@ -25,6 +25,11 @@
 
 const float mem_epsilon = 0.05;
 
+template<class T>
+inline T samples_to_rows(const T samples) {
+    return ((samples * (samples - 1)) >> 1);
+}
+
 // Checks bbits, sketchsize and k-mer lengths are identical in
 // all sketches
 // throws runtime_error if mismatches (should be ensured in passing
@@ -56,6 +61,62 @@ void printSlice(const SketchSlice& s) {
 	std::cerr << s.ref_size << std::endl;
 	std::cerr << s.query_offset << std::endl;
 	std::cerr << s.query_size << std::endl;
+}
+
+// Functions to join matrices when doing computation in chunks
+void longToSquareBlock(NumpyMatrix& coreSquare,
+                       NumpyMatrix& accessorySquare,
+                       const SketchSlice& sketch_subsample,
+                       NumpyMatrix& blockMat,
+                       const bool self,
+                       const unsigned int num_threads) {
+    // Convert each long form column of Nx2 matrix into square distance matrix
+    // Add this square matrix into the correct submatrix (block) of the final square matrix
+    Eigen::VectorXf dummy_query_ref;
+    Eigen::VectorXf dummy_query_query;
+    
+    NumpyMatrix square_form;
+    if (self) {
+        square_form = long_to_square(blockMat.col(0), 
+                                    dummy_query_ref, 
+                                    dummy_query_query,
+                                    num_threads);
+    } else {
+        square_form = Eigen::Map<NumpyMatrix, 0, Eigen::InnerStride<2>>(blockMat.data(), 
+                                    sketch_subsample.ref_size, 
+                                    sketch_subsample.query_size);
+    }
+    coreSquare.block(sketch_subsample.ref_offset, 
+                        sketch_subsample.query_offset,
+                        sketch_subsample.ref_size, 
+                        sketch_subsample.query_size) = square_form; 
+
+    if (self) {
+        square_form = long_to_square(blockMat.col(1), 
+                                    dummy_query_ref, 
+                                    dummy_query_query,
+                                    num_threads);
+    } else {
+        square_form = Eigen::Map<NumpyMatrix, 0, Eigen::InnerStride<2>>(blockMat.data()+1, 
+                                            sketch_subsample.ref_size, 
+                                            sketch_subsample.query_size);
+
+    }
+    accessorySquare.block(sketch_subsample.ref_offset, 
+                        sketch_subsample.query_offset,
+                        sketch_subsample.ref_size, 
+                        sketch_subsample.query_size) = square_form; 
+}
+
+NumpyMatrix twoColumnSquareToLong(const NumpyMatrix& coreSquare,
+                       const NumpyMatrix& accessorySquare,
+                       const unsigned int num_threads) {
+    Eigen::VectorXf core_dists = square_to_long(coreSquare, num_threads);
+    Eigen::VectorXf accessory_dists = square_to_long(accessorySquare, num_threads);
+    
+    NumpyMatrix dists_ret_matrix(samples_to_rows(coreSquare.rows()), 2);
+    dists_ret_matrix << core_dists, accessory_dists; // Join columns 
+    return(dists_ret_matrix);
 }
 
 // Main function callable via API
@@ -116,7 +177,7 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
 	SketchSlice sketch_subsample;
 	unsigned int chunks = 1;
 	std::vector<float> dist_results(dist_rows * 2);
-	NumpyMatrix coreSquare, accessorySquare, dists_ret_matrix;
+	NumpyMatrix coreSquare, accessorySquare;
 	if (self)
 	{
 		// To prevent memory being exceeded, total distance matrix is split up into
@@ -124,7 +185,7 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
 		// as a square distance matrix. The i = j chunks are 'self', i < j can be skipped
 		// as they contain only lower triangle values, i > j work as query vs ref
 		chunks = floor(est_size / (mem_free * (1 - mem_epsilon))) + 1;
-		// chunks = 2;
+		chunks = 2;
         size_t calc_per_chunk = n_samples / chunks;
 		unsigned int num_big_chunks = n_samples % chunks;
 
@@ -152,10 +213,9 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
 				}
 				printSlice(sketch_subsample);
 				
-				std::vector<float> block_results;
                 if (chunk_i == chunk_j) {
 					// 'self' blocks
-					block_results = dispatchDists(
+					dist_results = dispatchDists(
 						ref_sketches,
 						ref_sketches,
 						ref_strides,
@@ -165,7 +225,7 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
 						true);
 				} else {
 					// 'query' block
-					block_results = dispatchDists(
+					dist_results = dispatchDists(
 						ref_sketches,
 						query_sketches,
 						ref_strides,
@@ -182,21 +242,22 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
                     // Copy results from device into Nx2 matrix
                     NumpyMatrix blockMat = \
                         Eigen::Map<Eigen::Matrix<float,Eigen::Dynamic,2,Eigen::ColMajor> > \
-                            (block_results.data(),block_results.size()/2,2);
+                            (dist_results.data(),dist_results.size()/2,2);
+                    printf("Mapped\n");
                     
                     // Convert each long form column of Nx2 matrix into square distance matrix
                     // Add this square matrix into the correct submatrix (block) of the final square matrix
+                    if (self) {
+
+                    }
                     longToSquareBlock(coreSquare,
                                         accessorySquare,
                                         sketch_subsample,
-                                        block_results,
+                                        blockMat,
+                                        chunk_i == chunk_j,
                                         num_cpu_threads);
-				} else {
-                    dists_ret_matrix = \
-                        Eigen::Map<Eigen::Matrix<float,Eigen::Dynamic,2,Eigen::ColMajor> > \
-                            (block_results.data(),block_results.size()/2,2);
-                }
-
+                    printf("To square\n");
+				} 
 			}
 			sketch_subsample.query_offset += sketch_subsample.query_size; 
 		}
@@ -206,25 +267,29 @@ NumpyMatrix query_db_cuda(std::vector<Reference>& ref_sketches,
 	{
 		sketch_subsample.ref_size = ref_sketches.size();
 		sketch_subsample.query_size = query_sketches.size();
-		std::vector<float> dist_results = dispatchDists(
-                                            ref_sketches,
-                                            query_sketches,
-                                            ref_strides,
-                                            query_strides,
-                                            sketch_subsample,
-                                            kmer_lengths,
-                                            false);	
-        dists_ret_matrix = \
-            Eigen::Map<Eigen::Matrix<float,Eigen::Dynamic,2,Eigen::ColMajor> >(dist_results.data(),dist_results.size()/2,2);
+        sketch_subsample.ref_offset = 0;
+        sketch_subsample.query_offset = 0;
+
+		dist_results = dispatchDists(ref_sketches,
+                                    query_sketches,
+                                    ref_strides,
+                                    query_strides,
+                                    sketch_subsample,
+                                    kmer_lengths,
+                                    false);	
 	}
 	
-	if (self && chunks > 1) {
+	NumpyMatrix dists_ret_matrix;
+    if (self && chunks > 1) {
 		// Chunked computation yields square matrix, which needs to be converted back to long
 		// form
 		dists_ret_matrix = twoColumnSquareToLong(coreSquare,
 												 accessorySquare,
 												 num_cpu_threads);
-	} 
+	} else {
+        dists_ret_matrix = \
+            Eigen::Map<Eigen::Matrix<float,Eigen::Dynamic,2,Eigen::ColMajor> >(dist_results.data(),dist_results.size()/2,2);
+    } 
 
 	return dists_ret_matrix;
 }
