+++ src/matrix_ops.cpp
@@ -36,11 +36,6 @@ inline size_t rows_to_samples(const T& longMat) {
     return 0.5*(1 + sqrt(1 + 8*(longMat.rows())));
 }
 
-template<class T>
-inline T samples_to_rows(const T samples) {
-    return ((samples * (samples - 1)) >> 1);
-}
-
 // These are inlined partially to avoid conflicting with the cuda
 // versions which have the same prototype
 inline long calc_row_idx(const long long k, const long n) {
@@ -346,48 +341,3 @@ Eigen::VectorXf square_to_long(const NumpyMatrix& squareDists,
 
     return(longDists);
 }
-
-// Used by the code in dist.cu
-#ifdef GPU_AVAILABLE
-void longToSquareBlock(NumpyMatrix& coreSquare,
-                       NumpyMatrix& accessorySquare,
-                       const SketchSlice& sketch_subsample,
-                       std::vector<float>& block_results,
-                       const unsigned int num_threads) {
-    NumpyMatrix blockMat = \
-        Eigen::Map<Eigen::Matrix<float,Eigen::Dynamic,2,Eigen::RowMajor> >(block_results.data(),block_results.size()/2,2);
-    
-    // Convert each long form column of Nx2 matrix into square distance matrix
-    // Add this square matrix into the correct submatrix (block) of the final square matrix
-    Eigen::VectorXf dummy_query_ref;
-    Eigen::VectorXf dummy_query_query;
-    NumpyMatrix long_form = long_to_square(blockMat.col(0), 
-                                            dummy_query_ref, 
-                                            dummy_query_query,
-                                            num_threads);
-    coreSquare.block(sketch_subsample.ref_offset, 
-                        sketch_subsample.query_offset,
-                        sketch_subsample.ref_size, 
-                        sketch_subsample.query_size) = long_form; 
-
-    long_form = long_to_square(blockMat.col(1), 
-                                dummy_query_ref, 
-                                dummy_query_query,
-                                num_threads);
-    accessorySquare.block(sketch_subsample.ref_offset, 
-                        sketch_subsample.query_offset,
-                        sketch_subsample.ref_size, 
-                        sketch_subsample.query_size) = blockMat.col(0); 
-}
-
-NumpyMatrix twoColumnSquareToLong(const NumpyMatrix& coreSquare,
-                       const NumpyMatrix& accessorySquare,
-                       const unsigned int num_threads) {
-    Eigen::VectorXf core_dists = square_to_long(coreSquare, num_threads);
-    Eigen::VectorXf accessory_dists = square_to_long(accessorySquare, num_threads);
-    
-    NumpyMatrix dists_ret_matrix(samples_to_rows(coreSquare.rows()), 2);
-    dists_ret_matrix << core_dists, accessory_dists; // Join columns 
-    return(dists_ret_matrix);
-}
-#endif
